mini-program module rename.
