+++ src/MiniProgram/MiniProgram.php
@@ -34,7 +34,7 @@ use EasyWeChat\Support\Traits\PrefixedContainer;
  * @property \EasyWeChat\MiniProgram\Sns\Sns $sns
  * @property \EasyWeChat\MiniProgram\Stats\Stats $stats
  * @property \EasyWeChat\MiniProgram\Server\Guard $server
- * @property \EasyWeChat\MiniProgram\Notice\Notice $notice
+ * @property \EasyWeChat\MiniProgram\TemplateMessage\TemplateMessage $template_message
  * @property \EasyWeChat\MiniProgram\QRCode\QRCode $qrcode
  * @property \EasyWeChat\MiniProgram\Material\Temporary $material_temporary
  * @property \EasyWeChat\MiniProgram\CustomerService\CustomerService $customer_service
